Implemented the Lore Guy in hub
The full functionally of the lore guy has been implemented.
Including dialogue, showing only the enemies that has been defeated as options, having the text ready for each enemy and a picture at the end of the conversation with the lore guy. These changes can be seen in Hub 10-15 scene